Configure an alternative error color for some filetypes
This is getting a bit experimental, but it looks like it works.  A use
case is to have a faint color in files where you don't care about
trailing highlights that much (notes).  Or maybe files that require
trailing whitespace (although the filetype specific coloring should
handle that).

It can be used in 'gitcommit' buffers created with 'git commit --amend
--verbose' where you don't want to be distracted by the patch column but
still want to see whitespace errors in the message.