Sync highlighting on WinEnter
Every buffer has a variable that determines if bad whitespace should be
displayed or not.  The display will be synced on WinEnter.

I expect the common use case is to either permanently view them or not.
So this should work out fine in most cases.  A window-specific variable
would be initialized with the default on each split which is presumably
not what is wanted.